requirements.txt
PySide6>=6.5.0
psutil>=5.9.0
pynput>=1.7.6
Pillow>=9.0.0
numpy>=1.20.0
opencv-python>=4.5.0
pywin32>=306
# For stability, paddleocr version is locked.
paddleocr==2.7.3
